+++ cli/azd/pkg/project/service_config.go
@@ -22,8 +22,6 @@ type ServiceConfig struct {
 	Language ServiceLanguageKind `yaml:"language"`
 	// The output path for build artifacts
 	OutputPath string `yaml:"dist"`
-	// The infrastructure module path relative to the root infra folder to use for this project
-	Module string `yaml:"module"`
 	// The optional docker options
 	Docker DockerProjectOptions `yaml:"docker"`
 	// The optional K8S / AKS options