+++ cli/azd/pkg/tools/docker/docker.go
@@ -199,6 +199,7 @@ func (d *docker) CheckInstalled(ctx context.Context) error {
 	if err != nil {
 		return fmt.Errorf("checking %s version: %w", d.Name(), err)
 	}
+	log.Printf("docker version: %s", dockerRes)
 	supported, err := isSupportedDockerVersion(dockerRes)
 	if err != nil {
 		return err